requirements.txt
coverage==4.3.4
gitdb2==2.0.0
GitPython==2.1.3
smmap2==2.0.1
